refactor(ui): extract coordinator-admin/data/target-group.ts — localStorage + resolve helpers
